Skip decompression for file entries that has a unk of 0
Fixed an issue where certain drp not being fully extracted. 

This was caused by file entries with an unk variable of 0, the unk variable within the drp.Entries[] determines whether or not the file is zlib compressed.  

The program unexpectedly exits with code 0xc0000374 if it attempts to decompress a file with unk variable of 0. I've added a check to skip the decompression process if it detects the entry has an unk variable of 0.