Extractions and Repacker fix (#1)
* DRP entry address offset fix

Original method determines the offset by offsetting the reader.BaseStream.postion with the value of Fileend Offset , which will results in the wrong offset, causing the char buffer issue or any other unexpected errors.

This method offsets the BaseAddress of the current file with its Entry Size to find the BaseAddress of the next file.

* Skip decompression for file entries that has a unk of 0

Fixed an issue where certain drp not being fully extracted. 

This was caused by file entries with an unk variable of 0, the unk variable within the drp.Entries[] determines whether or not the file is zlib compressed.  

The program unexpectedly exits with code 0xc0000374 if it attempts to decompress a file with unk variable of 0. I've added a check to skip the decompression process if it detects the entry has an unk variable of 0.

* Skip compressing unk=0 file entries when repacking

Skip compression process with file entries that has a unk variable of 0

* Proper file count

Whoops, forgot to count the amount of subfiles.